+++ src/main/java/com/minecolonies/core/entity/ai/workers/crafting/AbstractEntityAICrafting.java
@@ -16,6 +16,7 @@ import com.minecolonies.api.equipment.ModEquipmentTypes;
 import com.minecolonies.api.util.BlockPosUtil;
 import com.minecolonies.api.util.InventoryUtils;
 import com.minecolonies.api.util.ItemStackUtils;
+import com.minecolonies.api.util.StatsUtil;
 import com.minecolonies.api.util.Tuple;
 import com.minecolonies.api.util.constant.ColonyConstants;
 import com.minecolonies.core.Network;
@@ -49,6 +50,8 @@ import static com.minecolonies.api.util.constant.CitizenConstants.*;
 import static com.minecolonies.api.util.constant.Constants.DEFAULT_SPEED;
 import static com.minecolonies.api.util.constant.StatisticsConstants.ITEMS_CRAFTED;
 import static com.minecolonies.core.util.WorkerUtil.hasTooManyExternalItemsInInv;
+
+import static com.minecolonies.api.util.constant.StatisticsConstants.ITEMS_CRAFTED_DETAIL;
 
 /**
  * Abstract class for the principal crafting AIs.
@@ -107,13 +110,20 @@ public abstract class AbstractEntityAICrafting<J extends AbstractJobCrafter<?, J
     }
 
     /**
-     * Provides a hook for implementing building-specific stats logic related to the crafting request.
-     * No-op hook. Override this in your subclass to customize it.
+     * Records the crafting request in the building's statistics.
+     * Override this in your subclass to change the description of the crafting stat.
+     * @param request the request to record.
      */
-    protected void recordCraftingBuildingStats(IRequest<?> request, IRecipeStorage recipe)
+    public void recordCraftingBuildingStats(IRequest<?> request, IRecipeStorage recipe)
     {
+        if (recipe == null) 
+        {
+            return;
+        }
 
+        StatsUtil.trackStatByName(building, ITEMS_CRAFTED_DETAIL, recipe.getPrimaryOutput().getDescriptionId(), recipe.getPrimaryOutput().getCount());
     }
+
 
     /**
      * Initialize the crafter job and add all his tasks.
