+++ src/main/java/com/minecolonies/core/entity/ai/workers/crafting/AbstractEntityAIRequestSmelter.java
@@ -15,6 +15,7 @@ import com.minecolonies.api.entity.ai.statemachine.states.AIBlockingEventType;
 import com.minecolonies.api.entity.ai.statemachine.states.IAIState;
 import com.minecolonies.api.util.InventoryUtils;
 import com.minecolonies.api.util.ItemStackUtils;
+import com.minecolonies.api.util.StatsUtil;
 import com.minecolonies.api.util.Tuple;
 import com.minecolonies.api.util.WorldUtil;
 import com.minecolonies.api.util.constant.translation.RequestSystemTranslationConstants;
@@ -48,6 +49,7 @@ import static com.minecolonies.api.util.constant.CitizenConstants.TICKS_20;
 import static com.minecolonies.api.util.constant.Constants.*;
 import static com.minecolonies.api.util.constant.TranslationConstants.BAKER_HAS_NO_FURNACES_MESSAGE;
 import static com.minecolonies.api.util.constant.TranslationConstants.FURNACE_USER_NO_FUEL;
+import static com.minecolonies.api.util.constant.StatisticsConstants.ITEMS_SMELTED_DETAIL;
 
 /**
  * Crafts furnace stone related block when needed.
@@ -879,11 +881,18 @@ public abstract class AbstractEntityAIRequestSmelter<J extends AbstractJobCrafte
     }
 
     /**
-     * Provides a hook for implementing building-specific stats logic related to the smelting request.
-     * No-op hook. Override this in your subclass to customize it.
+     * Records the smelting request in the building's statistics.
+     * Override this in your subclass to change the description of the smelting stat.
+     *
+     * @param cookedStack the item stack that has been smelted.
      */
     protected void recordSmeltingBuildingStats(ItemStack cookedStack)
     {
-
+        if (cookedStack == null) 
+        {
+            return;
+        }
+        
+        StatsUtil.trackStatByName(building, ITEMS_SMELTED_DETAIL, cookedStack.getDescriptionId(),cookedStack.getCount());
     }
 }
